tests/webgl/conformance/more/functions/isTestsBadArgs.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<!--

/*
** Copyright (c) 2012 The Khronos Group Inc.
**
** Permission is hereby granted, free of charge, to any person obtaining a
** copy of this software and/or associated documentation files (the
** "Materials"), to deal in the Materials without restriction, including
** without limitation the rights to use, copy, modify, merge, publish,
** distribute, sublicense, and/or sell copies of the Materials, and to
** permit persons to whom the Materials are furnished to do so, subject to
** the following conditions:
**
** The above copyright notice and this permission notice shall be included
** in all copies or substantial portions of the Materials.
**
** THE MATERIALS ARE PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
** EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
** MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
** IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
** CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
** TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
** MATERIALS OR THE USE OR OTHER DEALINGS IN THE MATERIALS.
*/

-->
<link rel="stylesheet" href="../../../resources/js-test-style.css"/>
<script src="../../../js/js-test-pre.js"></script>
<script src="../../../js/webgl-test-utils.js"></script>

</head>
<body>
<div id="description"></div>
<div id="console"></div>
<canvas id="canvas_element" width="1" height="1"></canvas>
<script type="application/javascript">

function runTest()
{
    shouldThrow("gl.isBuffer(gl.createFramebuffer())");
    shouldThrow("gl.isBuffer(gl.createProgram())");
    shouldThrow("gl.isBuffer(gl.createRenderbuffer())");
    shouldThrow("gl.isBuffer(gl.createShader(gl.VERTEX_SHADER))");
    shouldThrow("gl.isBuffer(gl.createTexture())");

    shouldThrow("gl.isFramebuffer(gl.createBuffer())");
    shouldThrow("gl.isFramebuffer(gl.createProgram())");
    shouldThrow("gl.isFramebuffer(gl.createRenderbuffer())");
    shouldThrow("gl.isFramebuffer(gl.createShader(gl.VERTEX_SHADER))");
    shouldThrow("gl.isFramebuffer(gl.createTexture())");

    shouldThrow("gl.isProgram(gl.createBuffer())");
    shouldThrow("gl.isProgram(gl.createFramebuffer())");
    shouldThrow("gl.isProgram(gl.createRenderbuffer())");
    shouldThrow("gl.isProgram(gl.createShader(gl.VERTEX_SHADER))");
    shouldThrow("gl.isProgram(gl.createTexture())");

    shouldThrow("gl.isRenderbuffer(gl.createBuffer())");
    shouldThrow("gl.isRenderbuffer(gl.createFramebuffer())");
    shouldThrow("gl.isRenderbuffer(gl.createProgram())");
    shouldThrow("gl.isRenderbuffer(gl.createShader(gl.VERTEX_SHADER))");
    shouldThrow("gl.isRenderbuffer(gl.createTexture())");

    shouldThrow("gl.isShader(gl.createBuffer())");
    shouldThrow("gl.isShader(gl.createFramebuffer())");
    shouldThrow("gl.isShader(gl.createProgram())");
    shouldThrow("gl.isShader(gl.createRenderbuffer())");
    shouldThrow("gl.isShader(gl.createTexture())");

    shouldThrow("gl.isTexture(gl.createBuffer())");
    shouldThrow("gl.isTexture(gl.createFramebuffer())");
    shouldThrow("gl.isTexture(gl.createProgram())");
    shouldThrow("gl.isTexture(gl.createRenderbuffer())");
    shouldThrow("gl.isTexture(gl.createShader(gl.VERTEX_SHADER))");

    shouldBe("gl.isBuffer(null)",            "false");
    shouldBe("gl.isBuffer(undefined)",       "false");

    shouldBe("gl.isFramebuffer(null)",       "false");
    shouldBe("gl.isFramebuffer(undefined)",  "false");

    shouldBe("gl.isProgram(null)",           "false");
    shouldBe("gl.isProgram(undefined)",      "false");

    shouldBe("gl.isRenderbuffer(null)",      "false");
    shouldBe("gl.isRenderbuffer(undefined)", "false");

    shouldBe("gl.isShader(null)",            "false");
    shouldBe("gl.isShader(undefined)",       "false");

    shouldBe("gl.isTexture(null)",           "false");
    shouldBe("gl.isTexture(undefined)",      "false");
}

description("Tests type checking for isX() functions");
var gl = WebGLTestUtils.create3DContext(document.getElementById("canvas_element"));
runTest();
var successfullyParsed = true;
</script>

<script src="../../../js/js-test-post.js"></script>
</body>
</html>
